add resource validation to fe - refactor by reorganising & chopping bigger components
